requirements.txt
ultralytics==8.3.0
opencv-python==4.11.0.86
numpy==1.26.4
PyYAML==6.0.3
